requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
black==23.7.0
boto3==1.29.4
botocore==1.32.4
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-framework-nested==0.0.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.7
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
protobuf==4.25.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.7.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.4
zope.interface==6.1
